% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_accuracy.R
\name{estimate_accuracy}
\alias{estimate_accuracy}
\title{Estimate Accuracy}
\usage{
estimate_accuracy(x)
}
\arguments{
\item{x}{Data}
}
\value{
Estimated accuracy
}
\description{
Estimate the probability of correct predictions. Note that this ignores the
mapping intended by the classifier and instead uses the optimal
prediction-mapping for each response.
}
\details{
Data can be input in three ways:
\itemize{
\item A data frame with variables "y" for true labels and "r" for
confidence-binned responses. "y" needs to contain values -1 and +1 while
r needs to be factor with levels ordered such that the first half of the
factor levels are predictions for -1 and the second half for
predictions +1. For example with 4 factor levels: 1: high confidence
prediction for y=-1, 2: low confidence prediction for y=-1, 3: low
confidence prediction for y=+1, 4: high confidence prediction for y=+1.
\item A counts table with joint absolute frequencies. Rows correspond to true
labels (stimulus categories) and columns correspond to responses.
\item A contingency matrix with joint relative frequencies (as before but
normalized to sum up to 1).
}
}
