% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_classifier.R
\name{estimate_classifier}
\alias{estimate_classifier}
\title{Estimate Classifier}
\usage{
estimate_classifier(dat)
}
\arguments{
\item{dat}{Data frame with true labels representing stimulus categories
(variable "y") and confidence-binned responses (variable "r")}
}
\value{
Estimated classifier represented as a contingency table with true
labels as rows and responses as columns
}
\description{
Construct a contingency table as an estimate for the classifier's joint
probability distribution between true labels and responses.
}
