% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_meta_I.R
\name{estimate_meta_I (expressed in bits, i.e. log base is 2)}
\alias{estimate_meta_I (expressed in bits, i.e. log base is 2)}
\alias{estimate_meta_I}
\title{Estimate Meta-I}
\usage{
estimate_meta_I(x)
}
\value{
Meta-I value
}
\description{
Estimate Dayan (2023)'s information-theoretic measure of metacognitive accuracy meta-I.
Meta-I is the transmitted information minus the minimal information given the accuracy of the classifier,
$$meta-I = I(Y; \hat{Y}, C) - I(Y; \hat{Y})$$
}
\details{
This is equivalent to Dayan's formulation where meta-I is the information
that confidences transmit about the correctness of a response,
$$meta-I = I(Y = \hat{Y}; C).$$

Data can be input in three ways:
\itemize{
\item A data frame with variables "y" for true labels and "r" for
confidence-binned responses. "y" needs to contain values -1 and +1 while
r needs to be a factor with ordered levels such that the first half of
the levels correspond to predictions for y=-1 and the second half to
predictions for y=+1.
\item A counts table with joint absolute frequencies. Rows correspond to true
labels (stimulus categories) and columns correspond to responses.
\item A contingency matrix with joint relative frequencies (as before but
normalized to sum up to 1).
}
}
\references{
Dayan, P. (2023). Metacognitive Information Theory. Open Mind, 7, 392–411. \url{https://doi.org/10.1162/opmi_a_00091}
}
\author{
Sascha Meyen
}
