% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_meta_Ir1_acc.R
\name{estimate_meta_Ir1_acc}
\alias{estimate_meta_Ir1_acc}
\title{Estimate Meta Ir1 via Accuracy}
\usage{
estimate_meta_Ir1_acc(x, ...)
}
\value{
Meta-I^r_1 value
}
\description{
Estimate Dayan (2023)'s information-theoretic measure of metacognitive sensitivity, meta-I^r_1.
This is meta-I (see estimate_meta_I()) normalized by the meta-I that would be
produced by normal noise with the estimated sensitivity,
}
\details{
$$meta-Ir1 = meta-I / meta-I(d')$$.

In contrast to estimate_meta_Ir1, the sensitivity is here computed by
determining the accuracy and converting it to a sensitivity assuming an
unbiased observer, d' = 2*qnorm(accuracy).

Data can be input in three ways:
\itemize{
\item A data frame with variables "y" for true labels and "r" for
confidence-binned responses. "y" needs to contain values -1 and +1 while
r needs to be a factor with ordered levels such that the first half of
the levels correspond to predictions for y=-1 and the second half to
predictions for y=+1.
\item A counts table with joint absolute frequencies. Rows correspond to true
labels (stimulus categories) and columns correspond to responses.
\item A contingency matrix with joint relative frequencies (as before but
normalized to sum up to 1).
}
}
