% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_meta_Ir2.R
\name{estimate_meta_Ir2}
\alias{estimate_meta_Ir2}
\title{Estimate Meta Ir2}
\usage{
estimate_meta_Ir2(x)
}
\value{
Meta-I^r_2 value
}
\description{
Estimate Dayan (2023)'s meta-I^r_1. This is meta-I (see
estimate_meta_I()) normalized by the remaining uncertainty
about whether the predictions are correct or not, $H(Y = \hat{Y})$,
}
\details{
$$meta-Ir2 = meta-I / H(Y = \hat{Y}).$$

Data can be input in three ways:
\itemize{
\item A data frame with variables "y" for true labels and "r" for
confidence-binned responses. "y" needs to contain values -1 and +1 while
r needs to be a factor with ordered levels such that the first half of
the levels correspond to predictions for y=-1 and the second half to
predictions for y=+1.
\item A counts table with joint absolute frequencies. Rows correspond to true
labels (stimulus categories) and columns correspond to responses.
\item A contingency matrix with joint relative frequencies (as before but
normalized to sum up to 1).
}
}
\references{
Dayan, P. (2023). Metacognitive Information Theory. Open Mind, 7, 392–411. \url{https://doi.org/10.1162/opmi_a_00091}
}
\author{
Sascha Meyen
}
