% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sensitivity.R
\name{estimate_sensitivity}
\alias{estimate_sensitivity}
\title{Estimate Sensitivity d'}
\usage{
estimate_sensitivity(x)
}
\arguments{
\item{x}{Data}
}
\value{
Sensitivity value
}
\description{
Estimate regular sensitivity from SDT. Here, we assume that the first half
of the confidence-binned responses equals to predictions for the first
label (Y = +1) and the second half of the responses to predictions for the
second label (Y = -1).
}
\details{
Data can be input in three ways:
\itemize{
\item A data frame with variables "y" for true labels and "r" for
confidence-binned responses. "y" needs to contain values -1 and +1 while
r needs to be a factor with ordered levels such that the first half of
the levels correspond to predictions for y=-1 and the second half to
predictions for y=+1.
\item A counts table with joint absolute frequencies. Rows correspond to true
labels (stimulus categories) and columns correspond to responses.
\item A contingency matrix with joint relative frequencies (as before but
normalized to sum up to 1).
}
}
