% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitConf.R
\name{fitConf}
\alias{fitConf}
\title{Fit a static confidence model to data}
\usage{
fitConf(data, model)
}
\arguments{
\item{data}{data.frame containing the data to fit the model to. Should contain
following columns: stimulus, correct, and rating}

\item{models}{character of length 1. Model to be fit.
Implemented models: 'WEV', 'SDT', 'Noisy', 'PDA', and '2Chan'.}
}
\value{
data.frame with one row and columns for the fitted parameters and
additional information (BIC, AIC,...)
}
\description{
Fit a static confidence model to data
}
\examples{
#ToDo
}
