% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitConfModel.R
\name{fitConfModel}
\alias{fitConfModel}
\title{Fit one or more static confidence models to the data}
\usage{
fitConfModel(data)
}
\arguments{
\item{data}{data.frame containing the data to fit the model to. Should contain
following columns: stimulus, correct, rating, and diffCond}

\item{models}{character vector. Models to be fit.
Implemented models: 'WEV', 'SDT', 'Noisy', 'PDA', and '2Chan'. Want to implement 'lognorm', 'ITGc' and 'ITGcm'}
}
\value{
data.frame with one row for each fitted model and columns for the fitted parameters and
as well as model fit indicates (BIC, AICc, AIC)
}
\description{
Fit one or more static confidence models to the data
}
\examples{
#To-DO
}
