% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRespConfModels.R
\name{fitMultiRespConf}
\alias{fitMultiRespConf}
\title{fit several static confidence models to multiple participants}
\usage{
fitMultiRespConf(data, models, .parallel = FALSE, n.cores = NULL)
}
\arguments{
\item{data}{data.frame containing the data to fit the models to. Should contain
following columns: participant, stimulus, correct, and rating}

\item{models}{character. Vector of models to be fit for each participant.
Implemented models: 'WEV', 'SDT', 'Noisy', 'PDA', and '2Chan'.}

\item{.parallel}{logical. Whether to parallelise the fitting over models and participant}

\item{n.cores}{integer. Number of cores used for parallelisation. If NULL, the available
number of cores -1 will be used.}
}
\value{
data.frame with one row per combination of participant and model and
columns for participant, model, the fitted parameters and additional information
about the fit (BIC, AIC,...)
}
\description{
fit several static confidence models to multiple participants
}
\examples{
#ToDo

}
