% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accuracy.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Compute accuracy of a classifier}
\usage{
get_accuracy(classifier)
}
\arguments{
\item{classifier}{A contingency matrix with joint probabilities. Rows
correspond to true labels (stimulus categories) and columns correspond to
responses.}
}
\value{
Accuracy of the classifier
}
\description{
The accuracy is the sum of highest entries for each response in the
classifier's contingency matrix.
}
