% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytic_information_bounds.R
\name{get_analytic_binary_information_bounds}
\alias{get_analytic_binary_information_bounds}
\title{Get Analytic Binary Information Bounds}
\usage{
get_analytic_binary_information_bounds(prior, accuracies)
}
\arguments{
\item{prior}{Prior probability as a vector of two values that sums to 1}

\item{accuracies}{Accuracies of classifiers}
}
\value{
Data frame with highest and lowest possible transmitted information
}
\description{
Given the accuracies of classifiers predicting binary true labels with the
given prior, compute the highest and lowest possible transmitted
information of them. Information is given in bit (log base is 2).
}
