% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binary_entropy.R
\name{get_binary_entropy}
\alias{get_binary_entropy}
\title{Get Binary Entropy}
\usage{
get_binary_entropy(p)
}
\arguments{
\item{p}{Probability of one realiziation}
}
\value{
Entropy of the binary random variable
}
\description{
Compute entropy of a binary random variable where one realization has
probability p. Which of the two realiziations of a binary variable doesn't
matter because of symmetry, H2(p) = H2(1-p).
}
