% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts_table.R
\name{get_counts_table}
\alias{get_counts_table}
\title{Get Counts Table}
\usage{
get_counts_table(dat)
}
\arguments{
\item{dat}{Data frame with true labels representing stimulus categories
(variable "y") and confidence-binned responses (variable "r")}
}
\value{
Matrix with contingency table of absolute frequencies with true
labels as rows and responses as columns
}
\description{
Construct a contingency table with absolute frequencies for all
combinations of true labels and responses.
}
