% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Get Entropy}
\usage{
get_entropy(p, base = 2)
}
\arguments{
\item{p}{Probability vector, has to sum to 1}

\item{base}{Base for the logarithm. If base = 2, the entropy is measured in
bit. If base = e, the entropy is measured in nat. Default is base = 2.}
}
\value{
Entropy H(p)
}
\description{
Compute entropy of a discrete random variable with probability vector p.
}
