% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_information}
\alias{get_information}
\title{Compute transmitted information of a classifier}
\usage{
get_information(classifier)
}
\arguments{
\item{classifier}{A contingency matrix with joint probabilities. Rows
correspond to true labels (stimulus categories) and columns correspond to
responses.}
}
\value{
Transmitted information of the classifier's output about the label
}
\description{
Compute the transmitted information for a given classifier.
}
