% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normal_noise_classifier.R
\name{get_normal_noise_classifier}
\alias{get_normal_noise_classifier}
\title{Get Normal Noise Classifier}
\usage{
get_normal_noise_classifier(sensitivity = NULL, accuracy = NULL)
}
\arguments{
\item{sensitivity}{Shift of the normal distributions}

\item{accuracy}{Resulting classifier accuracy}
}
\value{
Normal noise classifier
}
\description{
Define a classifier with underlying normal distributions on observations.
The normal distribution is of equal variance (= 1) and the means are
shifted by the sensitivity.
}
\details{
Either accuracy or sensitivity has to be specified to determine the
classifier. If both are specified (not advised), accuracies are
prioritizied.
}
