% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normal_noise_information.R
\name{get_normal_noise_information}
\alias{get_normal_noise_information}
\title{Get Normal Noise Information}
\usage{
get_normal_noise_information(accuracies = NULL, sensitivities = NULL)
}
\arguments{
\item{accuracies}{Vector of accuracies}

\item{sensitivities}{Vector of sensitivities}
}
\value{
Transmitted information of a normal noise classifier
}
\description{
Determine the transmitted information of a normal noise classifier. Either
accuracies or sensitivities have to be specified to determine the
corresponding transmitted information. If both are specified (not advised
due to ambiguity), accuracies are prioritizied.
}
