% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pointwise_binary_transmitted_information.R
\name{get_pointwise_binary_transmitted_information}
\alias{get_pointwise_binary_transmitted_information}
\title{Binary Transmitted Mutual Information (PTI)}
\usage{
get_pointwise_binary_transmitted_information(c, p = 0.5)
}
\arguments{
\item{c}{Confidence in the predicted label}

\item{p}{Base rate}
}
\value{
Transmitted Information between binary random variable and prediction
}
\description{
Mutual information of a binary random variable (typically with equal base
rates, p = 0.5) and a prediction with confidence c.
}
