% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pointwise_transmitted_information.R
\name{get_pointwise_transmitted_information}
\alias{get_pointwise_transmitted_information}
\title{Pointwise Transmitted Information (PTI)}
\usage{
get_pointwise_transmitted_information(c, prior)
}
\arguments{
\item{c}{Confidence vector, is assumed to be perfectly calibrated}

\item{p}{Base rate}
}
\value{
Transmitted Information between random variable and prediction
}
\description{
Transmitted information (often called mutual information) of a random
variable and response with confidence vector c.
}
