% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RMI.R
\name{get_RMI}
\alias{get_RMI}
\title{Get Relative Meta-Information}
\usage{
get_RMI(classifier)
}
\arguments{
\item{classifier}{A contingency matrix with joint probabilities. Rows
correspond to true labels (stimulus categories) and columns correspond to
responses.}
}
\value{
Relative meta-information value
}
\description{
Compute the meta-information as a relative measure in the possible range
given the accuracy: Highest information for a given accuracy produces
RMI = 1, lowest information for a given accuracy produces RMI = 0.
}
