% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupBMS.R
\name{groupBMS}
\alias{groupBMS}
\title{Group-level Bayesian Model Comparison}
\usage{
groupBMS(fits, measure = "BIC", opts = list())
}
\arguments{
\item{fits}{a data frame as returned by \code{\link{fitConfModels}}.
Should contain a column \code{model}indicating the model name, a column
\code{participant}, indicating the grouping
structure of the data (e.g. participant IDs), and a column with the name
given by the \code{measure} argument containing the values of
the information criterion that should be used to approximate model evidence.}

\item{measure}{the name of the the information criterion
to approximate model evidence. The available options are 'BIC', 'AIC', and 'AICc'.}

\item{opts}{a list with options for the Variational Bayes algorithm to estimate
the parameter of the Dirichlet distribution. Following values can  be provided:
\itemize{
\item \code{maxiter} the maximum number of iterations (Default: 200)
\item \code{tol} the tolerance for changes in the free energy approximation
to stop the algorithm, if abs(FE(i+1)-FE(i))<tol the algorithm
is stopped (Default: 1e-4)
\item \code{eps} The number to substitute values of 0 in calls to log (Default: 1e-32)
}}
}
\value{
\code{groupBMS} returns a list with two entries:
\itemize{
\item \code{model_weights}: a dataframe with rows for each model and following columns:
\itemize{
\item \code{alpha}, the alpha parameter of the Dirichlet posterior over model probabilities in the population according to the random effects model,
\item \code{r}, the mean probabilities of each model in the population according to the random effects model,
\item \code{ep}, the exceedance probability for each model,
\item \code{pep}, the protected exceedance probabilities for each model, and
\item \code{fx_prop}, the posterior model probabilities if a fixed true model is assumed in the population).
}
\item \code{summary_stats}: a vector giving statistics for the Bayesian model comparison analyses:
\itemize{
\item \code{bor_re_null}, Bayesian omnibus risk for the comparison of the random effects model against the null model,
\item \code{bor_fixed_null}, Bayesian omnibus risk for the comparison of the fixed effects model against the null model,
\item \code{bor_re_fixed}, Bayesian omnibus risk for the comparison of the random effects model against the fixed effects model
\item \code{FE_re}, estimated Free Energy according to the random effects model,
\item \code{FE_fixed} Free Energy according to the fixed effects model,
\item \code{FE_null} Free Energy according to the null model.
}
}
}
\description{
\code{groupBMS} performs a Bayesian model comparison on the group level based on marginal
likelihoods (alias model evidence) obtained from model fits on the level of single subjects.
It compares three models of the distribution of model probabilities (see Daunizeau et al., 2014; Rigoux at al., 2014; Stephan et al., 2009):
\itemize{
\item The random effects model treats the generative model as random effect that could differ between subjects. Model probabilities are assumed to follow a Dirichelet distribution in the population.
\item The fixed effects model, according to which one single model generated the data of all subjects, but subjects can still different from each other due to different model parameters.
\item The null model, according to which  model frequencies are a priori assumed to be fixed to 1/K, where K is the number of models to be compared.
\code{groupBMS} can be used with the output of \code{\link{fitConfModels}},
i.e. a data frame with information criteria for different models and subjects,
using one of the information criteria to approximate the model evidence.
}
}
\details{
The function \code{groupBMS} can be used for Bayesian model selection on
the group level following the approach of Rigoux et al. (2014).
It therefore aggregates the available subject-level model evidences
according to a random effects model, which assumes that model probabilities
are distributed in the population according to a Dirichlet distribution.
The parameters of the Dirichelet distribution are determined using a variational Bayes algorithm.
\code{groupBMS} provides three statistics associated with the random effects model:
\itemize{
\item estimated model frequencies in the population,
\item The exceedance probability  represents the probability that a specific
model is more probable than any other model (Stephan et al., 2009).
\item The Protected exceedance probability represents the probability that a specific model
is more probable than any other model for a random subject out of the population
scaled by the Bayesian omnibus risk of assuming a random effects model over
a null effect model, in which all models have the same probability (Rigoux et al., 2014).
The function is an R implementation of the VBA_groupBMC function of the VBA toolbox for MATLAB (Daunizeau, et al., 2014).
}
}
\examples{
# Define a data frame with information criteria from model fits
# this is a sub-data.frame from an output of fitConfModels with
# 5 subjects, two models and rounded information criteria)
fitted_par <- data.frame(
  participant = rep(1:5, each=2),
  model = rep(c("WEV", "ITGcm"), 5),
  BIC = c(5360, 5550, 3773, 3963, 3441, 3503, 2613, 2706, 3566, 3695),
  AIC = c(5274, 5469, 3686, 3883, 3355, 3422, 2527, 2625, 3479, 3614),
  AICc = c(5275, 5469, 3687, 3883, 3356, 3422, 2527, 2625, 3480, 3614))
# Conduct group-level Bayesian model selection based on BIC
ModelComp <- groupBMS(fitted_par, measure="BIC")
ModelComp

}
\references{
Rigoux, L., Stephan, K. E., Friston, K. J., & Daunizeau, J. (2014). Bayesian model selection for group studies - revisited. \emph{NeuroImage}, 84, 971–985. doi: 10.1016/j.neuroimage.2013.08.065

Daunizeau, J., Adam, V., & Rigoux, L. (2014). Vba: A probabilistic treatment of nonlinear models for neurobiological and behavioural data. \emph{PLOS Computational Biology}, 10(1), e1003441. doi: 10.1371/journal.pcbi.1003441

Stephan, K. E., Penny, W. D., Daunizeau, J., Moran, R. J., & Friston, K. J. (2009). Bayesian model selection for group studies. Neuroimage, 46(4), 1004-1017.
}
\author{
Sebastian Hellmann, \email{sebastian.hellmann@ku.de}\cr
Manuel Rausch, \email{manuel.rausch@ku.de}
}
