% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_information.R
\name{normalize_information}
\alias{normalize_information}
\title{Normalize Information}
\usage{
normalize_information(prior, accuracy, information)
}
\arguments{
\item{prior}{Prior of the true labels}

\item{accuracy}{Probability of a correct prediction (with optimal coding)}

\item{information}{Transmitted information}
}
\value{
Numeric representing the normalized information (RMI)
}
\description{
Given the prior and accuracy, there are limits to the transmitted
information. Normalize the given information by these bounds.
}
