% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfModelFit.R
\name{plotConfModelFit}
\alias{plotConfModelFit}
\title{Plot the prediction of fitted parameters of one model of confidence over the corresponding real data}
\usage{
plotConfModelFit(data, fitted_pars, model = NULL)
}
\arguments{
\item{data}{a \code{data.frame} where each row is one trial, containing following
variables:
\itemize{
\item \code{diffCond} (optional; different levels of discriminability,
should be a factor with levels ordered from hardest to easiest),
\item \code{rating} (discrete confidence judgments, should be a factor with levels
ordered from lowest confidence to highest confidence;
otherwise will be transformed to factor with a warning),
\item \code{stimulus} (stimulus category in a binary choice task,
should be a factor with two levels, otherwise it will be transformed to
a factor with a warning),
\item \code{correct} (encoding whether the response was correct; should  be 0 for
incorrect responses and 1 for correct responses)
\item \code{participant} (some group ID, most often a participant identifier;
the models given in the second argument are fitted to each subset of \code{data}
determined by the different values of this column)
}}

\item{fitted_pars}{a \code{data.frame} with one row for each participant and model parameters in different columns.
fitted_pars also may contain a column called  \code{model}specifying the model to be visualized.
If there is no model column in data or if there are multiple models in fitted_pars,
it is necessary to specify the model argument.}

\item{model}{\code{character}. See \code{\link{fitConf}} for all available models}
}
\value{
a \code{data.frame} with one row for each combination of model and
participant. There are different columns for the model, the participant ID, and one
one column for each estimated model parameter (parameters
not present in a specific model are filled with NAs)

#' @examples1. Select two subjects from the masked orientation discrimination experiment

data <- subset(MaskOri, participant \%in\% c(1:2))
head(data)2. Fit some models to each subject of the masked orientation discrimination experiment

\donttest{
  # Fitting several models to several subjects takes quite some time
  # (about 10 minutes per model fit per participant on a 2.8GHz processor
  # with the default values of nInits and nRestart).
  # If you want to fit more than just two subjects,
  # we strongly recommend setting .parallel=TRUE
  Fits <- fitConfModels(data, models = "ITGc", .parallel = FALSE)
}3. Plot the predicted probabilies based on model and fitted parameter over the observed relative frequencies.

\donttest{
  # Fitting several models to several subjects takes quite some time
  # (about 10 minutes per model fit per participant on a 2.8GHz processor
  # with the default values of nInits and nRestart).
  # If you want to fit more than just two subjects,
  # we strongly recommend setting .parallel=TRUE
  myPlottedFit <- plotConfModelFit(data, Fits)
  myPlottedFit
}
}
\description{
The \code{plotConfModelFit} function plots the predicted distribution of discrimination responses
and confidence ratings created from a \code{data.frame} of parameters obtaind from \code{\link{fitConfModels}}
and overlays the predicted distribution over the data to which the model parameters were fitted.
}
