% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_meta_I.R
\name{estimate_meta_I}
\alias{estimate_meta_I}
\title{Estimate Meta-I}
\usage{
estimate_meta_I(x)
}
\value{
Meta-I value (expressed in bits, i.e. log base is 2)
}
\description{
Estimate meta-I, an information-theoretic measure of metacognitive sensitviity
proposed by Dayan (2023).
}
\details{
Meta-I is the mutual information between the confidence and accuracy and is calculated as
the transmitted information minus the minimal information given the accuracy,
/deqn{meta-I = I(Y; \hat{Y}, C) - I(Y; \hat{Y})}. This is equivalent to Dayan's formulation where meta-I is the information
that confidences transmit about the correctness of a response, /deqn{meta-I = I(Y = \hat{Y}; C).}
}
\references{
Dayan, P. (2023). Metacognitive Information Theory. Open Mind, 7, 392–411. \url{https://doi.org/10.1162/opmi_a_00091}
}
\author{
Sascha Meyen
}
