% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_RMI.R
\name{estimate_RMI}
\alias{estimate_RMI}
\title{Estimate Relative Meta-Information}
\usage{
estimate_RMI(x)
}
\arguments{
\item{x}{Data}
}
\value{
Relative meta-information value
}
\description{
Estimate relative meta-information, RMI. This is meta-I (see
estimate_meta_I()) normalized by the possible range of meta-I values
for the estimated accuracy,
}
\details{
$$RMI = meta-I / max meta-I.$$

Data can be input in three ways:
\itemize{
\item A data frame with variables "y" for true labels and "r" for
confidence-binned responses. "y" needs to contain values -1 and +1 while
r needs to be a factor with ordered levels such that the first half of
the levels correspond to predictions for y=-1 and the second half to
predictions for y=+1.
\item A counts table with joint absolute frequencies. Rows correspond to true
labels (stimulus categories) and columns correspond to responses.
\item A contingency matrix with joint relative frequencies (as before but
normalized to sum up to 1).
}
}
\references{
Dayan, P. (2023). Metacognitive Information Theory. Open Mind, 7, 392–411. \url{https://doi.org/10.1162/opmi_a_00091}
}
\author{
Sascha Meyen
}
